EPMRPP-91774 || moved util from editNotificationModal to modal
